refactor: change zip ops structure
